refactor(mobile): addressbook
