+++ apps/mobile/src/features/AddressBook/Contact/ContactName.container.tsx
@@ -1,6 +1,6 @@
 import { useAppSelector } from '@/src/store/hooks'
 import { selectContactByAddress } from '@/src/store/addressBookSlice'
-import { ContactName } from '@/src/features/AddressBook/components/Contact/ContactName'
+import { ContactName } from './components/ContactName'
 import { type TextProps } from 'tamagui'
 
 type Props = {
@@ -8,7 +8,7 @@ type Props = {
   textProps?: Partial<TextProps>
 }
 
-export const ContactContainer = ({ address, textProps }: Props) => {
+export const ContactNameContainer = ({ address, textProps }: Props) => {
   const contact = useAppSelector(selectContactByAddress(address))
   return <ContactName name={contact?.name} address={address} textProps={textProps} />
 }
