+++ apps/mobile/src/features/AddressBook/List/components/List/index.ts
@@ -1,3 +1,3 @@
 import { AddressBookList } from './AddressBookList'
 export { AddressBookList }
-export { AddressBookListContainer } from './AddressBookList.container'
+export { AddressBookListContainer } from '../../AddressBookList.container'
