--- apps/mobile/src/features/AddressBook/components/Contact/index.tsx
@@ -1,10 +1,0 @@
-export { ContactView } from './ContactView'
-
-// Modular components
-export {
-  ContactActionButton,
-  ContactAddressField,
-  ContactHeader,
-  ContactNameField,
-  ContactNetworkRow,
-} from './components'
